moving the template outside of the main package
