+++ pipeline_template/setup.cfg
