+++ pipeline_template/setup.py
@@ -21,7 +21,7 @@ conf.read(['setup.cfg'])
 
 
 # MODIFY THE NAME OF THE PACKAGE to be the one chosen
-NAME = 'my_pipeline'
+NAME = 'template'
 VERSION = '0.0'
 RELEASE = 'dev' not in VERSION
 
@@ -29,7 +29,7 @@ scripts = []
 
 entry_points = {
     'console_scripts': [
-        "template_script = my_pipeline.scripts.template_script:main"
+        "template_script = template.scripts.template_script:main"
     ]
 }
 
@@ -43,7 +43,7 @@ setup(name=NAME,
       long_description=open('README.txt').read(),
       author='you',
       author_email='you@gmail.com',
-      packages=['my_pipeline',],
+      packages=['template',],
       scripts=scripts,
       entry_points=entry_points
       )
