+++ pipeline_template/template/configs/framework.cfg
@@ -9,7 +9,7 @@
 # The name of the configuration.
 # 
 [DEFAULT]
-name = 'DRP-Example'
+name = 'Template'
 
 #
 # Data set monitoring interval.
@@ -40,12 +40,12 @@ print_trace = True
 # Directories must be listed in PYTHONPATH.
 # This is only to avoid explicit imports in the code.
 #
-pipeline_path = "", "my_pipeline.pipelines"
+pipeline_path = "", "template.pipelines"
 
 
 # Where to find primitives.
 #
-primitive_path = "my_pipeline.primitives", ""
+primitive_path = "template.primitives", ""
 
 #
 # File type for Data_set
