+++ pipeline_template/template/configs/logger.cfg
