+++ pipeline_template/template/configs/template.cfg
